+++ tests/TestCase.php
@@ -4,33 +4,14 @@ declare(strict_types=1);
 
 namespace MongoDB\Laravel\Tests;
 
-use Illuminate\Auth\Passwords\PasswordResetServiceProvider as BasePasswordResetServiceProviderAlias;
 use Illuminate\Foundation\Application;
-use MongoDB\Laravel\Auth\PasswordResetServiceProvider;
-use MongoDB\Laravel\MongoDBQueueServiceProvider;
 use MongoDB\Laravel\MongoDBServiceProvider;
 use MongoDB\Laravel\Tests\Models\User;
 use MongoDB\Laravel\Validation\ValidationServiceProvider;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 
-use function array_search;
-
 class TestCase extends OrchestraTestCase
 {
-    /**
-     * Get application providers.
-     *
-     * @param  Application $app
-     */
-    protected function getApplicationProviders($app): array
-    {
-        $providers = parent::getApplicationProviders($app);
-
-        unset($providers[array_search(BasePasswordResetServiceProviderAlias::class, $providers)]);
-
-        return $providers;
-    }
-
     /**
      * Get package providers.
      *
@@ -40,8 +21,6 @@ class TestCase extends OrchestraTestCase
     {
         return [
             MongoDBServiceProvider::class,
-            MongoDBQueueServiceProvider::class,
-            PasswordResetServiceProvider::class,
             ValidationServiceProvider::class,
         ];
     }